Add isValidEcl method
Fixes #1 although only checks that the expression is parseable - ie valid ECL - it may be that the expression itself might be unsupported so the only way to really know is to use intersectEcl or expandEcl and check for an exception. This still will have a purpose in reducing resource demand during tests of user input, for example, as a first pass.